refactor(client): step 3 -- peel small families into mixins (stream, branches, tokens, query, workspaces, misc) (#520)
Pure movement. Extracts six endpoint families out of the KeboolaClient
monolith into per-family mixins, each `class _XMixin(_CoreClient)` so ty
statically resolves the plumbing (self._request, self._queue_request,
shared instance attrs) with no # type: ignore:

  client/stream.py      _StreamMixin      (7 methods)
  client/branches.py    _BranchesMixin    (7)
  client/tokens.py      _TokensMixin      (8)
  client/query.py       _QueryMixin       (6)
  client/workspaces.py  _WorkspacesMixin  (8)
  client/misc.py        _MiscMixin        (4: search/oauth/encrypt/sync-actions)

KeboolaClient now composes these six mixins over _CoreClient; every method
body, signature and docstring moved verbatim, and each module gets exactly
the imports its methods use.

Judgment call: _MiscMixin.get_oauth_url calls self.create_short_lived_token
(a _TokensMixin method). Across the mixin boundary ty cannot see it on the
bare _CoreClient base, so the method's `self` is annotated
`self: "KeboolaClient"` (via a TYPE_CHECKING import) -- the canonical typed-
mixin idiom. Runtime is unchanged: KeboolaClient inherits both mixins, so
the call resolves at run time exactly as before. This is the only cross-
family self-call among these six families.

_client.py: 3165 -> 2258 LOC. make check green: 4648 passed, 8 skipped. ty clean.
