+++ src/keboola_agent_cli/client/_client.py
@@ -27,36 +27,39 @@ from ..constants import (
     FILE_DOWNLOAD_TIMEOUT,
     FILE_UPLOAD_TIMEOUT,
     IMPORT_JOB_MAX_WAIT,
-    METADATA_NOT_FOUND,
-    OAUTH_HOST,
-    OAUTH_PATH,
-    OTLP_BUCKET_PREFIX,
-    QUERY_JOB_MAX_WAIT,
-    QUERY_JOB_POLL_INTERVAL,
-    QUERY_RESULTS_PAGE_SIZE,
     STORAGE_JOB_MAX_WAIT,
-    STREAM_DEFAULT_BRANCH,
     VALID_POLL_STRATEGIES,
 )
 from ..errors import ErrorCode, KeboolaApiError
-from ..models import TokenVerifyResponse
-from ..stream_client import StreamClient, provision_otlp_sinks, stream_task_source_id
 from ._core import _CoreClient
 from ._transfer import (
     _assert_safe_download_url,
     _build_abs_upload_url,
     _CloudDownloader,
     _extract_cloud_error_code,
-    _extract_query_job_error,
     _iter_poll_intervals,
     _IterBytesReader,
     _s3_signed_headers,
 )
+from .branches import _BranchesMixin
+from .misc import _MiscMixin
+from .query import _QueryMixin
+from .stream import _StreamMixin
+from .tokens import _TokensMixin
+from .workspaces import _WorkspacesMixin
 
 logger = logging.getLogger(__name__)
 
 
-class KeboolaClient(_CoreClient):
+class KeboolaClient(
+    _MiscMixin,
+    _StreamMixin,
+    _BranchesMixin,
+    _TokensMixin,
+    _QueryMixin,
+    _WorkspacesMixin,
+    _CoreClient,
+):
     """HTTP client for the Keboola Storage API and Queue API.
 
     Provides methods to interact with Keboola endpoints with built-in
@@ -65,490 +68,6 @@ class KeboolaClient(_CoreClient):
 
     Inherits _do_request() and _raise_api_error() from BaseHttpClient.
     """
-
-    def encrypt_values(
-        self,
-        project_id: int,
-        component_id: str,
-        data: dict[str, str],
-    ) -> dict[str, str]:
-        """Encrypt secret values via the Keboola Encryption API.
-
-        Sends a dict of {key: plaintext} and receives {key: encrypted}.
-        Keys must start with '#'. Encrypted values start with 'KBC::ProjectSecure::'.
-
-        Args:
-            project_id: Keboola project numeric ID.
-            component_id: Component identifier (e.g. 'keboola.ex-db-snowflake').
-            data: Dict of secret keys to encrypt (e.g. {'#password': 'my-secret'}).
-
-        Returns:
-            Dict of {key: encrypted_value}.
-        """
-        response = self._encrypt_request(
-            "POST",
-            "/encrypt",
-            params={"projectId": project_id, "componentId": component_id},
-            json=data,
-        )
-        return response.json()
-
-    def run_sync_action(
-        self,
-        component_id: str,
-        action: str,
-        config_data: dict[str, Any],
-        branch_id: int | None = None,
-        timeout: float | None = None,
-    ) -> Any:
-        """Run a synchronous component action via the Sync Actions API.
-
-        POSTs to ``/actions`` on the ``sync-actions.{stack-suffix}`` host.
-        Valid action names are component-defined (surfaced as
-        ``synchronous_actions`` in component metadata, e.g. ``testConnection``,
-        ``getTables``); the API validates them server-side.
-
-        Args:
-            component_id: Component identifier (e.g. 'keboola.ex-db-mysql').
-            action: Sync action name (freeform; component-defined).
-            config_data: The configData payload (typically
-                ``{"parameters": ..., "storage": ...}``). May carry secrets --
-                never log it.
-            branch_id: If set, sent as ``branchId``; omitted entirely for the
-                production branch (the API treats an absent key as default).
-            timeout: Optional per-request timeout in seconds (sync actions can
-                run long, e.g. ``getTables`` against a large database).
-
-        Returns:
-            The action result verbatim (opaque dict or list; shape is
-            action-specific).
-        """
-        body: dict[str, Any] = {
-            "configData": config_data,
-            "componentId": component_id,
-            "action": action,
-        }
-        if branch_id is not None:
-            body["branchId"] = branch_id
-        request_kwargs: dict[str, Any] = {"json": body}
-        if timeout is not None:
-            request_kwargs["timeout"] = timeout
-        response = self._sync_actions_request("POST", "/actions", **request_kwargs)
-        return response.json()
-
-    def verify_token(self) -> TokenVerifyResponse:
-        """Verify the storage API token and retrieve project information.
-
-        Returns:
-            TokenVerifyResponse with project name, ID, and token description.
-
-        Raises:
-            KeboolaApiError: If token is invalid (401) or other API error.
-        """
-        response = self._request("GET", "/v2/storage/tokens/verify")
-        data = response.json()
-
-        owner = data.get("owner", {})
-        # /v2/storage/tokens/verify carries `organization` at the TOP level
-        # (NOT nested under `owner` like I'd previously assumed -- three
-        # rounds of broken backfill traced back to this mismatch). The
-        # payload is minimal -- only `{"id": "73"}` on the GCP us-east4
-        # stack -- so org name has to come from the Manage API path.
-        org = data.get("organization") or {}
-        org_id_raw = org.get("id")
-        # Storage API serializes org id as a string ("73"); normalise to int
-        # so callers and persisted ProjectConfig.org_id can keep its int
-        # type without each consumer doing the cast.
-        org_id: int | None
-        try:
-            org_id = int(org_id_raw) if org_id_raw is not None else None
-        except (TypeError, ValueError):
-            org_id = None
-        response = TokenVerifyResponse(
-            token_id=str(data.get("id", "")),
-            token_description=data.get("description", ""),
-            project_id=owner.get("id"),
-            project_name=owner.get("name", ""),
-            owner_name=owner.get("name", ""),
-            default_backend=owner.get("defaultBackend", "snowflake"),
-            features=owner.get("features", []),
-            org_id=org_id,
-            # Top-level `organization` block does NOT carry a name; that
-            # field is Manage-API-only. Leave None and let the UI show
-            # the id (e.g. "#73") as a fallback until `org setup` fills
-            # in the human-readable name.
-            org_name=None,
-        )
-        # Refresh the features cache on every successful verify so explicit
-        # callers stay consistent with the cached view used by has_feature().
-        self._features_cache = frozenset(response.features)
-        return response
-
-    def get_project_info(self) -> dict[str, Any]:
-        """Return full project/token info from /v2/storage/tokens/verify.
-
-        Unlike verify_token() which parses only a subset of fields into
-        TokenVerifyResponse, this method returns the complete raw API response
-        so callers can access all fields (features, limits, metrics, etc.).
-
-        Returns:
-            Full JSON response dict from /v2/storage/tokens/verify.
-
-        Raises:
-            KeboolaApiError: If token is invalid (401) or other API error.
-        """
-        response = self._request("GET", "/v2/storage/tokens/verify")
-        return response.json()
-
-    def create_short_lived_token(
-        self,
-        description: str,
-        component_access: list[str],
-        expires_in: int = 3600,
-    ) -> dict[str, Any]:
-        """Create a short-lived Storage API token restricted to a component.
-
-        POST /v2/storage/tokens
-
-        Args:
-            description: Human-readable token description.
-            component_access: List of component IDs this token may access.
-            expires_in: Token lifetime in seconds (default: 3600 = 1 hour).
-
-        Returns:
-            Token dict from the API, including the 'token' field.
-        """
-        response = self._request(
-            "POST",
-            "/v2/storage/tokens",
-            data={
-                "description": description,
-                "expiresIn": str(expires_in),
-                "componentAccess[]": component_access,
-            },
-        )
-        return response.json()
-
-    def create_scoped_token(
-        self,
-        *,
-        description: str,
-        bucket_permissions: dict[str, str] | None = None,
-        component_access: list[str] | None = None,
-        can_read_all_file_uploads: bool = False,
-        expires_in: int | None = None,
-    ) -> dict[str, Any]:
-        """Create a scoped Storage API token (``POST /v2/storage/tokens``).
-
-        The general form of :meth:`create_short_lived_token`: instead of only a
-        component allow-list it also expresses **bucket** permissions, so a
-        caller can mint the narrow "upload Files + write one sink bucket,
-        expiring, nothing else" token a capture device needs (Keboola's
-        single-bucket-write pattern).
-
-        Note on Files upload: a Files upload (``POST /v2/storage/files/prepare``)
-        is a generic Storage write available to any valid Storage token -- it is
-        **not** gated by ``componentAccess`` or ``canReadAllFileUploads``. Grant
-        ``bucket_permissions={sink_bucket: "write"}`` for the sink write;
-        ``can_read_all_file_uploads`` only widens *reading* files uploaded by
-        *other* tokens (a device sees its own uploads regardless).
-
-        The acting token must carry ``canManageTokens`` (the API rejects the
-        create otherwise -- surfaced as an ``ACCESS_DENIED`` :class:`KeboolaApiError`
-        with the token masked). The returned dict is the raw API response; its
-        ``token`` field is a **one-time** secret reveal -- persist only ``id``
-        (for :meth:`delete_token` / :meth:`refresh_token`) and ``expires``.
-
-        Args:
-            description: Human-readable token description (per-device label).
-            bucket_permissions: ``{bucketId: "read" | "write"}`` grants.
-            component_access: Component IDs the token may run (often empty for a
-                capture device that only uploads Files + streams OTLP).
-            can_read_all_file_uploads: If True the token may read files uploaded
-                by other tokens (default False = only its own uploads).
-            expires_in: Lifetime in seconds; ``None`` = never expires.
-        """
-        data: dict[str, Any] = {"description": description}
-        if expires_in is not None:
-            data["expiresIn"] = str(expires_in)
-        if can_read_all_file_uploads:
-            # Storage API reads this form field as truthy; omit it (=> default
-            # false) rather than sending "0" so an unset scope stays minimal.
-            data["canReadAllFileUploads"] = "1"
-        for bucket_id, permission in (bucket_permissions or {}).items():
-            data[f"bucketPermissions[{bucket_id}]"] = permission
-        if component_access:
-            data["componentAccess[]"] = component_access
-        response = self._request("POST", "/v2/storage/tokens", data=data)
-        return response.json()
-
-    def delete_token(self, token_id: str) -> None:
-        """Revoke a Storage API token immediately (``DELETE /v2/storage/tokens/{id}``).
-
-        Returns 204 and the token stops authenticating at once. Only a
-        **non-master** token can be deleted (the API refuses to delete the master
-        token). Use this for active per-device revocation instead of waiting for
-        the token to expire.
-        """
-        self._request("DELETE", f"/v2/storage/tokens/{quote(str(token_id), safe='')}")
-
-    def refresh_token(self, token_id: str) -> dict[str, Any]:
-        """Rotate a Storage API token (``POST /v2/storage/tokens/{id}/refresh``).
-
-        Generates a **new** token value and returns the updated token dict; the
-        **old** token string becomes immediately invalid (rotation, not
-        additive), so every place using it must be updated. The token id is
-        stable across a refresh.
-        """
-        response = self._request(
-            "POST", f"/v2/storage/tokens/{quote(str(token_id), safe='')}/refresh"
-        )
-        return response.json()
-
-    # ------------------------------------------------------------------
-    # Data Streams (per-device OTLP sources) -- delegate to StreamClient
-    # ------------------------------------------------------------------
-
-    def _get_stream_client(self) -> StreamClient:
-        """Lazily build (and cache) a :class:`StreamClient` over this stack+token.
-
-        The Stream control plane lives on a sibling host (``stream.<region>``)
-        and authenticates with the same Storage token, so it is reachable from
-        the same ``(stack_url, token)`` this client already holds.
-        """
-        if self._stream_client is None:
-            self._stream_client = StreamClient(stack_url=self._stack_url, token=self._token)
-        return self._stream_client
-
-    @staticmethod
-    def _stream_source_detail(
-        source: dict[str, Any], branch_id: str, sink_bucket_id: str | None
-    ) -> dict[str, Any]:
-        """Normalise a raw Stream ``source`` object into the returned detail dict.
-
-        Flattens the OTLP block so the caller gets ``otlp_url`` (with the ingest
-        secret embedded, **unmasked** -- the lib layer hands it to the device
-        once and never persists it) without digging into ``source["otlp"]``. The
-        raw ``source`` is echoed under ``source`` so nothing is lost.
-        """
-        otlp = source.get("otlp") or {}
-        source_id = source.get("sourceId", "")
-        return {
-            "id": source_id,
-            "source_id": source_id,
-            "name": source.get("name", ""),
-            "type": source.get("type", ""),
-            "description": source.get("description", ""),
-            "branch_id": branch_id,
-            "otlp_url": otlp.get("url", ""),
-            "otlp_secret": otlp.get("secret", ""),
-            "base_endpoint": otlp.get("baseUrl", ""),
-            "sink_bucket_id": sink_bucket_id,
-            "source": source,
-        }
-
-    def create_stream_source(
-        self,
-        name: str,
-        *,
-        source_type: str = "otlp",
-        description: str = "",
-        branch_id: str = STREAM_DEFAULT_BRANCH,
-        provision_sinks: bool = True,
-    ) -> dict[str, Any]:
-        """Create a per-device OTLP (or HTTP) stream source and return its detail.
-
-        Async under the hood: the Stream API returns a 202 ``Task`` which is
-        polled to completion here (the caller sees one blocking call). For an
-        ``otlp`` source with ``provision_sinks`` (default) the logs/metrics/traces
-        sinks and the ``in.c-otlp-<sourceId>`` sink bucket are auto-created so
-        OTLP data actually lands in Storage -- and so a scoped device token can be
-        granted ``write`` on that bucket. Pass ``provision_sinks=False`` for a
-        bare source.
-
-        Per-device sources are the unit of isolated event-plane revocation:
-        delete one device's source (:meth:`delete_stream_source`) without
-        rotating a secret shared with other devices.
-
-        Returns a dict with the flattened endpoint: ``id`` / ``source_id``,
-        ``otlp_url`` (secret **unmasked**), ``otlp_secret``, ``sink_bucket_id``
-        (``None`` when no sinks were provisioned), and the raw ``source``.
-        Requires a Storage token privileged to manage Data Streams.
-        """
-        stream = self._get_stream_client()
-        task = stream.create_source(
-            branch_id, name=name, source_type=source_type, description=description or None
-        )
-        finished = stream.wait_for_task(task)
-        source_id = stream_task_source_id(finished) or name
-        sink_bucket_id: str | None = None
-        if provision_sinks and source_type == "otlp":
-            provision_otlp_sinks(stream, branch_id, source_id)
-            sink_bucket_id = f"{OTLP_BUCKET_PREFIX}{source_id}"
-        source = stream.get_source(branch_id, source_id)
-        return self._stream_source_detail(source, branch_id, sink_bucket_id)
-
-    @staticmethod
-    def _sink_bucket_from_sinks(sinks_raw: dict[str, Any]) -> str | None:
-        """Return the bucket of the source's first table sink, or None if it has none.
-
-        Derived from the ACTUAL sinks rather than assumed from the source type,
-        so a source created with ``provision_sinks=False`` (or outside kbagent)
-        truthfully reports ``None`` instead of a bucket that does not exist.
-        """
-        for sink in sinks_raw.get("sinks", []):
-            table_id = (sink.get("table") or {}).get("tableId", "")
-            if "." in table_id:
-                return table_id.rsplit(".", 1)[0]
-        return None
-
-    def get_stream_source(
-        self, source_id: str, *, branch_id: str = STREAM_DEFAULT_BRANCH
-    ) -> dict[str, Any]:
-        """Fetch one stream source's detail (endpoint + secret + sink bucket).
-
-        ``sink_bucket_id`` reflects the source's ACTUAL sinks (``None`` when it
-        has none -- e.g. created with ``provision_sinks=False`` or outside
-        kbagent), so a caller never scopes a device token to a bucket that does
-        not exist.
-        """
-        stream = self._get_stream_client()
-        source = stream.get_source(branch_id, source_id)
-        sinks = stream.list_sinks(branch_id, source.get("sourceId", source_id))
-        return self._stream_source_detail(source, branch_id, self._sink_bucket_from_sinks(sinks))
-
-    def list_stream_sources(
-        self, *, branch_id: str = STREAM_DEFAULT_BRANCH
-    ) -> list[dict[str, Any]]:
-        """List the project's stream sources (raw source objects; find-or-create by name)."""
-        raw = self._get_stream_client().list_sources(branch_id)
-        return list(raw.get("sources", []))
-
-    def delete_stream_source(
-        self, source_id: str, *, branch_id: str = STREAM_DEFAULT_BRANCH
-    ) -> None:
-        """Delete a stream source (per-device revocation) -- async task polled to completion."""
-        stream = self._get_stream_client()
-        task = stream.delete_source(branch_id, source_id)
-        stream.wait_for_task(task)
-
-    def global_search(
-        self,
-        query: str,
-        project_id: int,
-        types: list[str] | None = None,
-        branch_type: str = "production",
-        branch_id: int | None = None,
-        limit: int = 50,
-        offset: int = 0,
-        regex: bool = False,
-    ) -> dict[str, Any]:
-        """Search for items by name across the project using the Storage API global-search endpoint.
-
-        Calls GET /v2/storage/global-search with the given query and optional type filters.
-        This performs textual (name-based) search only — it does not scan configuration bodies.
-        Results are scoped to the single project identified by ``project_id``.
-
-        Args:
-            query: Search string to match against item names.
-            project_id: Numeric Keboola project ID (required by the API).
-            types: Optional list of item types to filter results. Supported values:
-                   ``bucket``, ``table``, ``flow``, ``transformation``, ``configuration``,
-                   ``configuration-row``, ``workspace``, ``shared-code``.
-                   If None or empty, all types are returned.
-            branch_type: ``"production"`` (default) or ``"development"``.
-            branch_id: Required when ``branch_type="development"``; ignored otherwise.
-            limit: Maximum number of results to return (default 50, max 100).
-            offset: Pagination offset (default 0).
-            regex: When True, run the query as a case-insensitive whole-term
-                   regular expression over entity names (Storage API
-                   ``mode=regex``). Omitted from the request otherwise.
-
-        Returns:
-            Raw API response dict with keys ``"all"`` (total count) and
-            ``"items"`` (list of matching item dicts).
-
-        Raises:
-            KeboolaApiError: On API errors (auth, network, rate limits).
-        """
-        params: dict[str, Any] = {
-            "query": query,
-            "projectIds[]": project_id,
-            "limit": limit,
-            "offset": offset,
-        }
-        if types:
-            params["types[]"] = types
-        if regex:
-            params["mode"] = "regex"
-        if branch_type == "development" and branch_id is not None:
-            params["branchTypes[]"] = "development"
-            params["branchIds[]"] = branch_id
-        else:
-            params["branchTypes[]"] = "production"
-
-        response = self._request("GET", "/v2/storage/global-search", params=params)
-        return response.json()
-
-    def get_oauth_url(
-        self,
-        component_id: str,
-        config_id: str,
-        redirect_url: str | None = None,
-    ) -> str:
-        """Generate an OAuth authorization URL for a component configuration.
-
-        Creates a short-lived, component-scoped Storage API token and builds
-        the URL the user must open to grant OAuth access.
-
-        Args:
-            component_id: The component ID (e.g. 'keboola.ex-google-drive').
-            config_id: The configuration ID to authorize.
-            redirect_url: Optional URL the OAuth wizard returns to after the
-                flow completes (passed as the ``returnUrl`` query param).
-
-        Returns:
-            The full OAuth authorization URL as a string.
-        """
-        from urllib.parse import urlencode, urlunsplit
-
-        token_response = self.create_short_lived_token(
-            description=f"Short-lived token for OAuth URL - {component_id}/{config_id}",
-            component_access=[component_id],
-            expires_in=3600,
-        )
-        sapi_token = token_response["token"]
-
-        query: dict[str, str] = {"token": sapi_token, "sapiUrl": self._stack_url}
-        if redirect_url:
-            query["returnUrl"] = redirect_url
-        query_params = urlencode(query)
-        fragment = f"/{component_id}/{config_id}"
-
-        return urlunsplit(("https", OAUTH_HOST, OAUTH_PATH, query_params, fragment))
-
-    def get_project_features(self) -> frozenset[str]:
-        """Return the project's feature flags, fetching once per client lifetime.
-
-        Calls ``verify_token()`` lazily on first request and caches the result.
-        Subsequent calls do not trigger HTTP. The cache lives for the life of
-        the ``KeboolaClient`` instance, which is one CLI invocation -- short
-        enough that staleness across feature toggles is not a practical risk.
-        """
-        if self._features_cache is None:
-            self.verify_token()
-        # _features_cache is non-None here: verify_token() always sets it (or
-        # raises on auth/network failure, which propagates to the caller).
-        assert self._features_cache is not None
-        return self._features_cache
-
-    def has_feature(self, feature: str) -> bool:
-        """True if the project owner has ``feature`` enabled.
-
-        Convenience wrapper over ``get_project_features()`` for code paths
-        that branch on a single flag (e.g. ``"storage-branches"``).
-        """
-        return feature in self.get_project_features()
 
     def list_components(
         self,
@@ -1024,141 +543,6 @@ class KeboolaClient(_CoreClient):
             "DELETE",
             f"{prefix}/components/{quote(component_id)}/configs/{quote(config_id)}/rows/{quote(row_id)}",
         )
-
-    def create_dev_branch(self, name: str, description: str = "") -> dict[str, Any]:
-        """Create a new development branch (waits for async job to complete).
-
-        The Storage API returns an async job. This method polls until the job
-        completes and returns the branch data from the job results.
-
-        Args:
-            name: Branch name.
-            description: Optional branch description.
-
-        Returns:
-            Branch dict with id, name, description, created, etc.
-
-        Raises:
-            KeboolaApiError: If the API call or job fails.
-        """
-        body: dict[str, str] = {"name": name}
-        if description:
-            body["description"] = description
-        response = self._request("POST", "/v2/storage/dev-branches", json=body)
-        job = self._wait_for_storage_job(response.json())
-        return job.get("results", {})
-
-    def delete_dev_branch(self, branch_id: int) -> None:
-        """Delete a development branch (waits for async job to complete).
-
-        Args:
-            branch_id: The branch ID to delete.
-
-        Raises:
-            KeboolaApiError: If the API call or job fails.
-        """
-        response = self._request("DELETE", f"/v2/storage/dev-branches/{branch_id}")
-        self._wait_for_storage_job(response.json())
-
-    def list_dev_branches(self) -> list[dict[str, Any]]:
-        """List development branches for the project.
-
-        Returns:
-            List of branch dicts from the API.
-        """
-        response = self._request("GET", "/v2/storage/dev-branches")
-        return response.json()
-
-    def list_branch_metadata(self, branch_id: int | str = "default") -> list[dict[str, Any]]:
-        """List metadata entries on a branch.
-
-        GET /v2/storage/branch/{id}/metadata
-
-        Args:
-            branch_id: Branch ID or the literal "default" for the main branch.
-
-        Returns:
-            List of metadata dicts with keys: id, key, value, provider, timestamp.
-        """
-        response = self._request("GET", f"/v2/storage/branch/{branch_id}/metadata")
-        return response.json()
-
-    def set_branch_metadata(
-        self,
-        entries: list[tuple[str, str]],
-        branch_id: int | str = "default",
-    ) -> list[dict[str, Any]]:
-        """Bulk-set metadata key/value pairs on a branch.
-
-        POST /v2/storage/branch/{id}/metadata
-
-        Keboola's endpoint expects PHP-style array indices in the
-        form-urlencoded body, e.g.::
-
-            metadata[0][key]=KBC.projectDescription
-            metadata[0][value]=My project
-
-        httpx's ``data=`` accepts a mapping of str -> str and URL-encodes it.
-        Since each ``metadata[i][...]`` key is unique per index, a plain dict
-        preserves both ordering (Python 3.7+) and Keboola's expected shape.
-
-        Args:
-            entries: Ordered list of ``(key, value)`` metadata tuples.
-            branch_id: Branch ID or the literal "default" for the main branch.
-
-        Returns:
-            List of metadata dicts created/updated by the API.
-        """
-        form: dict[str, str] = {}
-        for i, (key, value) in enumerate(entries):
-            form[f"metadata[{i}][key]"] = key
-            form[f"metadata[{i}][value]"] = value
-        response = self._request(
-            "POST",
-            f"/v2/storage/branch/{branch_id}/metadata",
-            data=form,
-        )
-        return response.json()
-
-    def delete_branch_metadata(
-        self,
-        metadata_id: int | str,
-        branch_id: int | str = "default",
-    ) -> None:
-        """Delete a single metadata entry on a branch by its numeric ID.
-
-        DELETE /v2/storage/branch/{id}/metadata/{metadataId}
-
-        Args:
-            metadata_id: ID of the metadata entry (from ``list_branch_metadata``).
-            branch_id: Branch ID or the literal "default" for the main branch.
-        """
-        self._request(
-            "DELETE",
-            f"/v2/storage/branch/{branch_id}/metadata/{metadata_id}",
-        )
-
-    def get_branch_metadata_value(
-        self,
-        key: str,
-        branch_id: int | str = "default",
-    ) -> str | None | object:
-        """Return the value for a single metadata key on a branch, or None if absent.
-
-        Convenience wrapper around ``list_branch_metadata`` that filters by key.
-
-        Args:
-            key: Metadata key to look up (e.g. "KBC.projectDescription").
-            branch_id: Branch ID or the literal "default" for the main branch.
-
-        Returns:
-            The string value if the key exists (may be None if the API stored null),
-            or ``METADATA_NOT_FOUND`` sentinel if the key is not present.
-        """
-        for entry in self.list_branch_metadata(branch_id=branch_id):
-            if entry.get("key") == key:
-                return entry.get("value")
-        return METADATA_NOT_FOUND
 
     def list_buckets(
         self, include: str | None = None, branch_id: int | None = None
@@ -2855,75 +2239,6 @@ class KeboolaClient(_CoreClient):
             retryable=True,
         )
 
-    # --- Workspace CRUD ---
-
-    def list_workspaces(self, branch_id: int | None = None) -> list[dict[str, Any]]:
-        """List all workspaces in the project."""
-        prefix = f"/v2/storage/branch/{branch_id}" if branch_id else "/v2/storage"
-        response = self._request("GET", f"{prefix}/workspaces")
-        return response.json()
-
-    def get_workspace(self, workspace_id: int, branch_id: int | None = None) -> dict[str, Any]:
-        """Get workspace details (note: password is NOT included)."""
-        prefix = f"/v2/storage/branch/{branch_id}" if branch_id else "/v2/storage"
-        response = self._request("GET", f"{prefix}/workspaces/{workspace_id}")
-        return response.json()
-
-    def delete_workspace(self, workspace_id: int, branch_id: int | None = None) -> None:
-        """Delete a workspace (synchronous)."""
-        prefix = f"/v2/storage/branch/{branch_id}" if branch_id else "/v2/storage"
-        self._request("DELETE", f"{prefix}/workspaces/{workspace_id}")
-
-    def reset_workspace_password(
-        self, workspace_id: int, branch_id: int | None = None
-    ) -> dict[str, Any]:
-        """Reset workspace password. Returns new password."""
-        prefix = f"/v2/storage/branch/{branch_id}" if branch_id else "/v2/storage"
-        response = self._request("POST", f"{prefix}/workspaces/{workspace_id}/password")
-        return response.json()
-
-    def create_sandbox_config(
-        self,
-        name: str,
-        description: str = "",
-        backend_size: str = "small",
-        branch_id: int | None = None,
-    ) -> dict[str, Any]:
-        """Create a keboola.sandboxes configuration.
-
-        This is needed to make workspaces visible in the Keboola UI.
-        The UI only shows workspaces tied to a sandboxes config.
-
-        Args:
-            name: Human-readable name for the workspace.
-            description: Optional description.
-            backend_size: Backend size (small, medium, large).
-            branch_id: Branch ID. If provided, creates config in that branch.
-
-        Returns:
-            Configuration dict with id, name, etc.
-        """
-        config = {
-            "parameters": {
-                "runtime": {"shared": False},
-                "storage": {"input": {"tables": []}, "output": {"tables": []}},
-                "parameters": {"id": "", "blocks": []},
-                "backendSize": backend_size,
-            },
-            "runtime": {"shared": False},
-        }
-        prefix = f"/v2/storage/branch/{branch_id}" if branch_id else "/v2/storage"
-        response = self._request(
-            "POST",
-            f"{prefix}/components/keboola.sandboxes/configs",
-            data={
-                "name": name,
-                "description": description,
-                "configuration": json.dumps(config),
-            },
-        )
-        return response.json()
-
     def delete_config(
         self, component_id: str, config_id: str, branch_id: int | None = None
     ) -> None:
@@ -2940,226 +2255,4 @@ class KeboolaClient(_CoreClient):
         self._request(
             "DELETE",
             f"{prefix}/components/{safe_component}/configs/{safe_config}",
-        )
-
-    def create_config_workspace(
-        self,
-        branch_id: int,
-        component_id: str,
-        config_id: str,
-        backend: str = "snowflake",
-        login_type: str | None = None,
-        public_key: str | None = None,
-    ) -> dict[str, Any]:
-        """Create a workspace tied to a specific configuration.
-
-        Args:
-            branch_id: Branch ID (use main branch ID for production).
-            component_id: Component ID (e.g. keboola.snowflake-transformation).
-            config_id: Configuration ID.
-            backend: Workspace backend.
-            login_type: Optional Storage API loginType. Omitted when None.
-            public_key: Optional public key for key-pair workspaces. Omitted when None.
-
-        Returns:
-            Workspace dict including connection credentials.
-        """
-        safe_component = quote(component_id, safe="")
-        safe_config = quote(config_id, safe="")
-        payload: dict[str, Any] = {"backend": backend}
-        if login_type is not None:
-            payload["loginType"] = login_type
-        if public_key is not None:
-            payload["publicKey"] = public_key
-
-        response = self._request(
-            "POST",
-            f"/v2/storage/branch/{branch_id}/components/{safe_component}/configs/{safe_config}/workspaces",
-            json=payload,
-        )
-        return response.json()
-
-    def list_config_workspaces(
-        self,
-        branch_id: int,
-        component_id: str,
-        config_id: str,
-    ) -> list[dict[str, Any]]:
-        """List workspaces tied to a specific configuration."""
-        safe_component = quote(component_id, safe="")
-        safe_config = quote(config_id, safe="")
-        response = self._request(
-            "GET",
-            f"/v2/storage/branch/{branch_id}/components/{safe_component}/configs/{safe_config}/workspaces",
-        )
-        return response.json()
-
-    def load_workspace_tables(
-        self,
-        workspace_id: int,
-        tables: list[dict[str, Any]],
-        branch_id: int | None = None,
-        preserve: bool = False,
-    ) -> dict[str, Any]:
-        """Load tables into a workspace (async operation).
-
-        Args:
-            workspace_id: Target workspace ID.
-            tables: List of table load definitions, each with at minimum:
-                - source: table ID (e.g. "in.c-bucket.table")
-                - destination: target table name in workspace
-            branch_id: Branch ID. Required for workspaces on dev branches.
-            preserve: If True, keep existing tables in the workspace. Default is False
-                (workspace is cleared before loading).
-
-        Returns:
-            Completed storage job dict (polls until done).
-
-        Raises:
-            KeboolaApiError: If the load job fails or times out.
-        """
-        prefix = f"/v2/storage/branch/{branch_id}" if branch_id else "/v2/storage"
-        body: dict[str, Any] = {"input": tables, "preserve": preserve}
-        response = self._request(
-            "POST",
-            f"{prefix}/workspaces/{workspace_id}/load",
-            json=body,
-        )
-        return self._wait_for_storage_job(response.json())
-
-    # --- Query Service ---
-
-    def submit_query(
-        self,
-        branch_id: int,
-        workspace_id: int,
-        statements: list[str],
-        transactional: bool = False,
-    ) -> dict[str, Any]:
-        """Submit SQL statements to the Query Service.
-
-        Args:
-            branch_id: Branch ID.
-            workspace_id: Workspace ID.
-            statements: List of SQL statements to execute.
-            transactional: Whether to wrap in a transaction.
-
-        Returns:
-            Query job dict with id and status.
-        """
-        body: dict[str, Any] = {
-            "statements": statements,
-            "transactional": transactional,
-        }
-        response = self._query_request(
-            "POST",
-            f"/api/v1/branches/{branch_id}/workspaces/{workspace_id}/queries",
-            json=body,
-        )
-        return response.json()
-
-    def get_query_job(self, query_job_id: str) -> dict[str, Any]:
-        """Get query job status."""
-        response = self._query_request("GET", f"/api/v1/queries/{query_job_id}")
-        return response.json()
-
-    def export_query_results(
-        self,
-        query_job_id: str,
-        statement_id: str,
-        file_type: str = "csv",
-    ) -> str:
-        """Export query results as CSV (or other format).
-
-        Returns:
-            Raw CSV string of query results.
-        """
-        response = self._query_request(
-            "GET",
-            f"/api/v1/queries/{query_job_id}/{statement_id}/export",
-            params={"fileType": file_type},
-        )
-        return response.text
-
-    def get_query_results(
-        self,
-        query_job_id: str,
-        statement_id: str,
-        offset: int = 0,
-        page_size: int = QUERY_RESULTS_PAGE_SIZE,
-    ) -> dict[str, Any]:
-        """Fetch a page of inline statement results from the Query Service.
-
-        Unlike :meth:`export_query_results`, which materializes a CSV file via the
-        warehouse UNLOAD path (slow), this reads the already-computed result set
-        inline as JSON -- much faster for interactive queries. The endpoint is
-        paginated; ``offset``/``page_size`` walk the result set.
-
-        Args:
-            query_job_id: The query job ID.
-            statement_id: The statement ID within the job.
-            offset: Row offset to start from (for pagination).
-            page_size: Maximum rows to return in this page.
-
-        Returns:
-            Raw QueryResult dict, e.g.::
-
-                {
-                    "status": "completed",
-                    "columns": [{"name": "id", "type": "INTEGER", "nullable": false}],
-                    "data": [[1, "a"], [2, "b"]],
-                    "numberOfRows": 2,
-                }
-        """
-        response = self._query_request(
-            "GET",
-            f"/api/v1/queries/{query_job_id}/{statement_id}/results",
-            params={"offset": offset, "pageSize": page_size},
-        )
-        return response.json()
-
-    def get_query_history(
-        self,
-        branch_id: int,
-        workspace_id: int,
-    ) -> dict[str, Any]:
-        """Get query history for a workspace."""
-        response = self._query_request(
-            "GET",
-            f"/api/v1/branches/{branch_id}/workspaces/{workspace_id}/queries",
-        )
-        return response.json()
-
-    def wait_for_query_job(self, query_job_id: str) -> dict[str, Any]:
-        """Poll a Query Service job until it reaches a terminal state.
-
-        Args:
-            query_job_id: The query job ID.
-
-        Returns:
-            Completed query job dict.
-
-        Raises:
-            KeboolaApiError: If the query fails or times out.
-        """
-        deadline = time.monotonic() + QUERY_JOB_MAX_WAIT
-        while time.monotonic() < deadline:
-            job = self.get_query_job(query_job_id)
-            status = job.get("status", "")
-            if status == "completed":
-                return job
-            if status in ("error", "failed"):
-                raise KeboolaApiError(
-                    message=f"Query job failed: {_extract_query_job_error(job)}",
-                    status_code=500,
-                    error_code=ErrorCode.QUERY_JOB_FAILED,
-                    retryable=False,
-                )
-            time.sleep(QUERY_JOB_POLL_INTERVAL)
-
-        raise KeboolaApiError(
-            message=f"Query job {query_job_id} did not complete within {QUERY_JOB_MAX_WAIT}s",
-            status_code=504,
-            error_code=ErrorCode.QUERY_JOB_TIMEOUT,
-            retryable=True,
         )
